+++ src/RhinoInside.Revit.External/DB/Extensions/Mesh.cs
@@ -26,16 +26,24 @@ namespace RhinoInside.Revit.External.DB.Extensions
     /// <remarks>
     /// Calculates the centroid of the mesh using an approximation, with an accuracy
     /// suitable for architectural purposes. This will correspond only with the center
-    /// of gravity if the mesh represents a homogeneous structure of a single material.
+    /// of gravity if the mesh is closed.
     /// </remarks>
     /// <param name="mesh"></param>
+    /// <param name="dimension"></param>
     /// <returns>The XYZ point of the Centroid of this mesh.</returns>
-    public static XYZ ComputeCentroid(this Mesh mesh)
+    public static XYZ ComputeCentroid(this Mesh mesh, int dimension)
     {
+      if (0 > dimension || dimension > 3)
+        throw new System.ArgumentOutOfRangeException(nameof(dimension));
+
+      if (dimension == 0)
+        return XYZExtension.ComputeMeanPoint(mesh.Vertices);
+
       Sum weights = default;
       Sum centroidX = default, centroidY = default, centroidZ = default;
-      var numTriangles = mesh.NumTriangles;
+      var factor = dimension + 1.0;
 
+      var numTriangles = mesh.NumTriangles;
       for (int t = 0; t < numTriangles; ++t)
       {
         var triangle = mesh.get_Triangle(t);
@@ -48,9 +56,24 @@ namespace RhinoInside.Revit.External.DB.Extensions
         vY.Add(v0.Y, v1.Y, v2.Y);
         vZ.Add(v0.Z, v1.Z, v2.Z);
 
-        var w = (v1 - v0).CrossProduct(v2 - v0, 0D).GetLength(0D);
-        weights.Add(w);
-        w /= 3.0;
+        var w = 0.0;
+        switch (dimension)
+        {
+          case 1:
+            w = (v1 - v0).GetLength() + (v2 - v1).GetLength() + (v0 - v2).GetLength();
+            break;
+
+          case 2:
+            w = (v1 - v0).CrossProduct(v2 - v0).GetLength();
+            break;
+
+          case 3:
+            w = v0.TripleProduct(v1, v2);
+            break;
+        }
+
+        weights += w;
+        w /= factor;
 
         centroidX.Add(vX.Value * w);
         centroidY.Add(vY.Value * w);
@@ -60,6 +83,18 @@ namespace RhinoInside.Revit.External.DB.Extensions
       var weightsSum = weights.Value;
       return new XYZ(centroidX.Value / weightsSum, centroidY.Value / weightsSum, centroidZ.Value / weightsSum);
     }
+
+    /// <summary>
+    /// Returns the Centroid of this mesh.
+    /// </summary>
+    /// <remarks>
+    /// Calculates the centroid of the mesh using an approximation, with an accuracy
+    /// suitable for architectural purposes. This will correspond only with the center
+    /// of gravity if the mesh represents a homogeneous structure of a single material.
+    /// </remarks>
+    /// <param name="mesh"></param>
+    /// <returns>The XYZ point of the Centroid of this mesh.</returns>
+    public static XYZ ComputeCentroid(this Mesh mesh) => ComputeCentroid(mesh, 3);
 
     /// <summary>
     /// Return the mean of all triangle normals